Update for 16w36a/snapshot, fixed server player rotation and cleanup.

+++ src/main/java/org/lanternpowered/server/network/vanilla/message/codec/play/CodecPlayOutPlayerJoinGame.java
@@ -47,7 +47,7 @@ public final class CodecPlayOutPlayerJoinGame implements Codec<MessagePlayOutPla
         }
         buf.writeByte(gameMode);
         buf.writeInteger(message.getDimensionType().getInternalId());
-        buf.writeByte(message.getDifficulty().getInternalId());
+        buf.writeByte((byte) message.getDifficulty().getInternalId());
         buf.writeByte((byte) Math.min(message.getPlayerListSize(), 255));
         buf.writeString("default"); // Not used
         buf.writeBoolean(message.getReducedDebug());
